Use environment variables to override parameters in PKGBUILD
The default value is now set via extended Bash substitution, where first
the value of the variable is checked, and if the user didn't set it, the
default value is used to assign it a default value, which is written in
the format:

_var=${_var=default}

This allows you to both set another variable value and unset the
variable. For example:

env _use_auto_optimization= _processor_opt=native_intel makepkg -sricCf

It is possible to use export as well.

All changes were made through the command:
find . -name "PKGBUILD" | xargs -I {} sed -i '1,178s/\(^_[A-Za-z_][^\n].*\)\(=.*\)/\1=\$\{\1\2\}/g' {}