6.1.10-2 && 6.2.rc7-2
Revert the recent refactor of the PKGBUILD. The refactor is currently breaking the compilation of the ZFS MODULE.
So far we did not find, why the recent changes are breaking it. We need to investigate further into this.

Anyways, we stay now at the old PKGBUILDs, which have the same functionality as the refactor. The ZFS Module is compiling correctly with it.

Signed-off-by: Peter Jung <admin@ptr1337.dev>