chore: Add $_use_gcc_suffix
Use `$_use_gcc_suffix` instead of `$_use_lto_suffix` since LTO kernel is going to be the default now.
The latter variable will be preserved for users that still want to use it.

I have modified the if statement with the logic:
1. If using LTO + enabled LTO suffix -> cachyos-lto
2. If no LTO + no kCFI + enabled GCC suffix -> cachyos-gcc
3. All other combinations -> cachyos

The main rationale behind (2) is that kCFI isn't LTO but it also isn't gcc.

I have also moved $pkgbase outside of the if else condition because it is the same regardless of condition

Signed-off-by: Eric Naim <dnaim@cachyos.org>