linux-cachyos-autofdo: Introduce Package
This kernel is intended to provide a preprofiled kernel, so that user can test this. This will follow the "linux-cachyos" kernel, and after a longer testing period this kernel will be dropped and we are optimizing the linux-cachyos kernel as default.

We will profile the kernel every 2 minor versions with following workloads (as soon LLVM 19 is out):
1. Kernel Compilation
2. cachyos-benchmarker
3. Maybe some other specific workloads, like svt-av1

The profiles needs to be done each architecure. For now we will make the profiles for:
- x86-64-v3
- Zen4

Since there are the most of our users.

Workflow:
1. Compile Kernel with _autofdo=y and _build_debug=y
2. Boot into kenrel in QEMU or host machine
3. Profile the kernel and convert the profile see blogpost
4. Put the profile into the sourcedir andn name the profile perf.afdo
5. Enable _autofdo_use

In the future we will also add propellor

Signed-off-by: Peter Jung <admin@ptr1337.dev>