linux-cachyos: Conditionally add clang-llvm to headers dependencies
Since we are compiling the default kernel with clang rather than GCC, we need to pull clang as a dependency
to ensure functionality with building out-of-tree kernel modules. For CachyOS users, this is already seamless
because our dkms package is dependent on clang, but AUR users and non-CachyOS users can suffer from this.

Note that this generates a bogus .SRCINFO when !_is_clang_kernel() but that doesn't matter for 100% of usecases
because it's default True (unless some third party repository defaults to GCC in which case they should fix it
on their own).

Closes #346

Signed-off-by: Eric Naim <dnaim@cachyos.org>