script-v3-v4: Handle $autofdo
In the PKGBUILD, we have a check whenever $autofdo is a non-zero string, it will build with clang. This breaks
the workflow of the script when building with GCC. It's not that obvious that it's broken because $pkgsuffix is
also broken.

We can fix this by forcing $autofdo off during GCC compilations and flipping it back on again when compiling with
ThinLTO

Signed-off-by: Eric Naim <dnaim@cachyos.org>