PKGBUILD: Make LTO a hard dependency when building with clang
In hindsight, there's not much to be gained by compiling clang without LTO. Faster compile times can be achieved
but the additional time it takes using LTO hardly matters on modern computers.

Forcing this on also makes it simple for the build script we use because the clang-specific features aren't
available on GCC (obviously).

Also refactored the pkgsuffix conditions to use _is_lto_kernel instead to make it cleaner.

Suggested-by: Vasiliy Stelmachenok <ventureo@cachyos.org>
Signed-off-by: Eric Naim <dnaim@cachyos.org>