PKGBUILD: Introduce rt $_preempt option
In an attempt to correctly classify $_cpusched as strictly CPU schedulers, remove both `rt` and `rt-bore` from $_cpusched
option and move it to $_preempt. In fact, this would be more correct and make more sense, since PREEMPT_RT only makes more
code preemptible compared to PREEMPT and doesn't do anything else special.

One of the blockers for this was sourcing the OOT patches that were needed for the i915 module to be compiled, but
this is not a problem as there is a conditional check for rt to not build zfs.

Signed-off-by: Eric Naim <dnaim@cachyos.org>