PKGBUILD: Remove global `hardened` option from $_cpusched
Like the default and handheld kernel, the hardened kernel differs from the other generic cachyos kernels in that it uses
its own special config[1]. This means that you can't reproduce a "proper" hardened kernel by simply choosing the `hardened`
$_cpusched option.

Instead, we should source the patch directly in the main source array, like what we currently do with the handheld patches.
This will be addressed in the coming 6.13 stable sync, where no functional changes are intended so even if hardened still
stays on 6.12, it should be fine.

[1] https://gitlab.archlinux.org/archlinux/packaging/packages/linux-hardened/-/blob/main/config

Signed-off-by: Eric Naim <dnaim@cachyos.org>