linux-cachyos-rc: Disable dynamic preemption
By observing a few results, it is observed that dynamic preemption can somehow blur the results between the various
preemption models. One example is [1] where there seems to be a 5% difference in throughput between voluntary vs full
with dynamic preemption disabled but then the difference goes to a measly 0.6% when dynamic preemption is enabled.
I've also been able to reproduce this behaviour across various benchmarks, one of them shared in [2].

Due to that, let's disable dynamic preemption for all kernel configurations (or atleast the default). Note that
PREEMPT_DYNAMIC is kept enabled for fully preempted kernels, as it is observed that it doesn't seem to matter for those,
so having more options is the priority in this case.

[1] https://lore.kernel.org/all/579b7ea34ef6e2f7c955abdfc0929fe1af36faef.camel@gmx.de/
[2] https://github.com/CachyOS/linux-cachyos/issues/339

Signed-off-by: Eric Naim <dnaim@cachyos.org>