Reduce number of kernel configurators
We have 4 interfaces for kernel configuration, most of which have the
same functionality and differ only in the rendering backend. gconfig
uses GTK2 which is EOL, and unlikely anyone will install it just for
gconfig when there are alternatives like xconfig which is based on Qt5 /
Qt 6. Similarly in the case of TUI, instead of menuconfig, here we have
ncurses based nconfig which is supported by most terminals.

Signed-off-by: Vasiliy Stelmachenok <ventureo@cachyos.org>