linux-cachyos-rc: Drop auto-cpu-optimization.sh
With the addition of CONFIG_X86_NATIVE_CPU, we can drop this external
script in favour of compiler autodetection. The only Kconfigs that are
included now are the 4 x86_64 ISA levels (contemplating only 3 but to
maintain compatibility with other kernels, keep it 4 for now) and
znver4.

Since the script is dropped, the default behaviour when empty will now
select CONFIG_X86_NATIVE_CPU, until all kernels match this behaviour,
then we can change this if desired.

Signed-off-by: Eric Naim <dnaim@cachyos.org>