linux-cachyos-deckify: Build msi-wmi-platform into the kernel
steamos-manager needs msi-wmi-platform to be loaded for TDP controls on
the MSI Claw to actually work. This can be done in one of three ways:
1. Implement autoloading functionality into the module itself for the
   Claw
2. Add `msi-wmi-platform` to /usr/lib/modules.load.d
3. Build it directly into the kernel

(1) is the best solution, but is the hardest to implement at a distro
level, and there isn't an estimate on when we can expect functionality
to be introduced. This leaves it to (2) and (3). I believe that (2) is
incredibly futile, since building it as a module first and
unconditionally loading it second is incredibly pointless.

There is also precedent in the Steam Deck modules being directly built,
so there's no reason why `msi-wmi-platform` can't be either.

Signed-off-by: Eric Naim <dnaim@cachyos.org>